+++ praudio/transforms/powerspectrogram.py
@@ -7,6 +7,7 @@ import logging
 
 import numpy as np
 
+from praudio.transforms.transform import TransformType, Transform
 from praudio.transforms.stft import STFT
 from praudio.io.signal import Signal
 
@@ -14,18 +15,13 @@ from praudio.io.signal import Signal
 logger = logging.getLogger(__name__)
 
 
-class PowerSpectrogram(STFT):
+class PowerSpectrogram(Transform):
     """This class extracts a power spectrogram from a signal. It's a
-    subclass of the STFT class.
+    concrete Transform.
 
     Attributes:
-        - frame_length: Length of the windowed signal after padding with zeros
-        - hop_length: Number of audio samples between adjacent STFT columns
-        - win_length: Each frame of audio is windowed by window of length
-            win_length and then padded with zeros to match frame_length
-        - window: Windowing method employed for STFT. Default is 'hann'
-        - power: Exponent to which power spectrogram is raised. Default
-            is 2
+        - power: Power to which we raise the spectrum. Default is 2
+        - stft: STFT object used to extract Short-Time Fourier Transform
     """
 
     def __init__(self,
@@ -34,23 +30,28 @@ class PowerSpectrogram(STFT):
                  win_length: int = 2048,
                  window: str = "hann",
                  power: int = 2):
-        super().__init__(frame_length, hop_length, win_length, window)
+        super().__init__(TransformType.POWERSPECTROGRAM)
         self.power = power
-        self.name = "power_spectrogram"
-        logger.info("Instantiated MagnitudeSpectroram object")
+        self.stft = STFT(frame_length,
+                         hop_length,
+                         win_length,
+                         window)
+
 
     def process(self, signal: Signal) -> Signal:
         """Extract power spectrogram from waveform and modify signal.
+        First we extract the STFT. Then, we apply the absolute value and
+        raise it to the target power.
 
         :param signal: Signal object. Note: this transforms works only with
             waveform data
 
         :return: Modified signal
         """
-        signal = super().process(signal)
+        signal = self.stft.process(signal)
+        signal.name = self._prepend_transform_name(signal.name)
         signal.data = self._raise_to_power(np.abs(signal.data))
-        signal.name = self.name
-        logger.info("Extracted power spectrogram for %s", signal.file)
+        logger.info("Applied %s to %s", self.name.value, signal.file)
         return signal
 
     def _raise_to_power(self, array: np.ndarray):