+++ praudio/transforms/scaling/rowstandardiser.py
@@ -3,6 +3,7 @@
 import numpy as np
 
 from praudio.transforms.scaling.scaler import Scaler
+from praudio.transforms.transform import TransformType
 from praudio.errors import NotA2DSignalError
 
 
@@ -20,7 +21,7 @@ class RowStandardiser(Scaler):
     """
 
     def __init__(self):
-        super().__init__("rowstandardiser")
+        super().__init__(TransformType.ROWSTANDARDISER)
         self._min_std = 1e-7
 
     def _scale(self, array: np.ndarray):