+++ praudio/transforms/stft.py
@@ -7,7 +7,7 @@ import logging
 
 import librosa
 
-from praudio.transforms.transform import Transform
+from praudio.transforms.transform import Transform, TransformType
 from praudio.io.signal import Signal
 
 
@@ -33,13 +33,13 @@ class STFT(Transform):
                  frame_length: int = 2048,
                  hop_length: int = 1024,
                  win_length: int = 2048,
-                 window: str = "hann"):
-        super().__init__("stft")
+                 window: str = "hann",
+                 name: TransformType = TransformType.STFT):
+        super().__init__(name)
         self.frame_length = frame_length
         self.hop_length = hop_length
         self.win_length = win_length
         self.window = window
-        logger.info("Instantiated STFT object")
 
     def process(self, signal: Signal) -> Signal:
         """Extract complex STFT from waveform and modify signal.
@@ -49,13 +49,11 @@ class STFT(Transform):
 
         :return: Modified signal
         """
-        stft = librosa.stft(signal.data,
-                            n_fft=self.frame_length,
-                            hop_length=self.hop_length,
-                            win_length=self.win_length,
-                            window=self.window)
-        signal.data = stft
-        signal.name = self.name
-        logger.info("Extracted Short-Time Fourier Transform for %s",
-                     signal.file)
+        signal.name = self._prepend_transform_name(signal.name)
+        signal.data = librosa.stft(signal.data,
+                                   n_fft=self.frame_length,
+                                   hop_length=self.hop_length,
+                                   win_length=self.win_length,
+                                   window=self.window)
+        logger.info("Applied %s to %s", self.name.value, signal.file)
         return signal