+++ praudio/transforms/transform.py
@@ -1,8 +1,31 @@
-"""This module provides an abstract interface for all transforms."""
+"""
+This module provides an abstract interface for all transforms (Transform)
+and an enumerations with all available transforms (TransformType).
+TransformType should be updated as new transforms are added.
+"""
 
+import logging
 from abc import ABC, abstractmethod
+from enum import Enum
 
 from praudio.io.signal import Signal
+
+
+logger = logging.getLogger(__name__)
+
+
+class TransformType(Enum):
+    """Enumeration class with all available transforms."""
+
+    LOG = "log"
+    MAGNITUDESPECTROGRAM = "magnitudespectrogram"
+    MINMAXSCALER = "minmaxscaler"
+    MELSPECTROGRAM = "melspectrogram"
+    MFCC = "mfcc"
+    POWERSPECTROGRAM = "powerspectrogram"
+    STFT = "stft"
+    ROWSTANDARDISER = "rowstandardiser"
+    STANDARDISER = "standardiser"
 
 
 class Transform(ABC):
@@ -14,8 +37,9 @@ class Transform(ABC):
         - name: The name of the transforms
     """
 
-    def __init__(self, name: str):
+    def __init__(self, name: TransformType):
         self.name = name
+        logger.info("Instantiated %s transform", self.name)
 
     @abstractmethod
     def process(self, signal: Signal) -> Signal:
@@ -26,3 +50,6 @@ class Transform(ABC):
 
         :return: New signal object with transformed values
         """
+
+    def _prepend_transform_name(self, string):
+        return self.name.value + "_" + string