+++ test/transforms/test_magnitudespectrogram.py
@@ -2,6 +2,7 @@ import numpy as np
 import pytest
 
 from praudio.transforms.magnitudespectrogram import MagnitudeSpectrogram
+from praudio.transforms.transform import TransformType
 from utils import sample_signal
 
 
@@ -13,11 +14,7 @@ def magnitude_spectrogram():
 def test_magnitude_spectrogram_instance_is_instantiated_correctly(
         magnitude_spectrogram):
     assert isinstance(magnitude_spectrogram, MagnitudeSpectrogram)
-    assert magnitude_spectrogram.frame_length == 100
-    assert magnitude_spectrogram.hop_length == 50
-    assert magnitude_spectrogram.win_length == 75
-    assert magnitude_spectrogram.window == "hann"
-    assert magnitude_spectrogram.name == "magnitude_spectrogram"
+    assert magnitude_spectrogram.name == TransformType.MAGNITUDESPECTROGRAM
 
 
 def test_magnitude_spectrogram_is_extracted(magnitude_spectrogram, sample_signal):
@@ -29,7 +26,7 @@ def test_magnitude_spectrogram_is_extracted(magnitude_spectrogram, sample_signal
     AND the modified Signal object with the new magnitude spectrogram is returned
     """
     signal = magnitude_spectrogram.process(sample_signal)
-    assert signal.name == "magnitude_spectrogram"
+    assert signal.name == "magnitudespectrogram_stft_dummy"
     assert type(signal.data) == np.ndarray
     assert len(signal.data.shape) == 2
     assert type(signal.data[0][0]) == np.float_